Safari-611.1.10/LayoutTests/svg/custom/js-svg-constructors-expected.txt

SVGAngle.SVG_ANGLETYPE_UNKNOWN PASSED
SVGAngle.SVG_ANGLETYPE_UNSPECIFIED PASSED
SVGAngle.SVG_ANGLETYPE_DEG PASSED
SVGAngle.SVG_ANGLETYPE_RAD PASSED
SVGAngle.SVG_ANGLETYPE_GRAD PASSED
SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_UNKNOWN PASSED
SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_IDENTITY PASSED
SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_TABLE PASSED
SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_DISCRETE PASSED
SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_LINEAR PASSED
SVGComponentTransferFunctionElement.SVG_FECOMPONENTTRANSFER_TYPE_GAMMA PASSED
SVGCSSRule.COLOR_PROFILE_RULE EXPECTED FAILURE (actual: 'undefined' expected: '7')
SVGFEBlendElement.SVG_FEBLEND_MODE_UNKNOWN PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_NORMAL PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_MULTIPLY PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_SCREEN PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_DARKEN PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_LIGHTEN PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_OVERLAY PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_COLOR_DODGE PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_COLOR_BURN PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_HARD_LIGHT PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_SOFT_LIGHT PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_DIFFERENCE PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_EXCLUSION PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_HUE PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_SATURATION PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_COLOR PASSED
SVGFEBlendElement.SVG_FEBLEND_MODE_LUMINOSITY PASSED
SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_UNKNOWN PASSED
SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_MATRIX PASSED
SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_SATURATE PASSED
SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_HUEROTATE PASSED
SVGFEColorMatrixElement.SVG_FECOLORMATRIX_TYPE_LUMINANCETOALPHA PASSED
SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_UNKNOWN PASSED
SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OVER PASSED
SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_IN PASSED
SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_OUT PASSED
SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_ATOP PASSED
SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_XOR PASSED
SVGFECompositeElement.SVG_FECOMPOSITE_OPERATOR_ARITHMETIC PASSED
SVGFEConvolveMatrixElement.SVG_EDGEMODE_UNKNOWN PASSED
SVGFEConvolveMatrixElement.SVG_EDGEMODE_DUPLICATE PASSED
SVGFEConvolveMatrixElement.SVG_EDGEMODE_WRAP PASSED
SVGFEConvolveMatrixElement.SVG_EDGEMODE_NONE PASSED
SVGFEDisplacementMapElement.SVG_CHANNEL_UNKNOWN PASSED
SVGFEDisplacementMapElement.SVG_CHANNEL_R PASSED
SVGFEDisplacementMapElement.SVG_CHANNEL_G PASSED
SVGFEDisplacementMapElement.SVG_CHANNEL_B PASSED
SVGFEDisplacementMapElement.SVG_CHANNEL_A PASSED
SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_UNKNOWN PASSED
SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_ERODE PASSED
SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE PASSED
SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_UNKNOWN PASSED
SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE PASSED
SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_TURBULENCE PASSED
SVGFETurbulenceElement.SVG_STITCHTYPE_UNKNOWN PASSED
SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH PASSED
SVGFETurbulenceElement.SVG_STITCHTYPE_NOSTITCH PASSED
SVGGradientElement.SVG_SPREADMETHOD_UNKNOWN PASSED
SVGGradientElement.SVG_SPREADMETHOD_PAD PASSED
SVGGradientElement.SVG_SPREADMETHOD_REFLECT PASSED
SVGGradientElement.SVG_SPREADMETHOD_REPEAT PASSED
SVGLength.SVG_LENGTHTYPE_UNKNOWN PASSED
SVGLength.SVG_LENGTHTYPE_NUMBER PASSED
SVGLength.SVG_LENGTHTYPE_PERCENTAGE PASSED
SVGLength.SVG_LENGTHTYPE_EMS PASSED
SVGLength.SVG_LENGTHTYPE_EXS PASSED
SVGLength.SVG_LENGTHTYPE_PX PASSED
SVGLength.SVG_LENGTHTYPE_CM PASSED
SVGLength.SVG_LENGTHTYPE_MM PASSED
SVGLength.SVG_LENGTHTYPE_IN PASSED
SVGLength.SVG_LENGTHTYPE_PT PASSED
SVGLength.SVG_LENGTHTYPE_PC PASSED
SVGMarkerElement.SVG_MARKERUNITS_UNKNOWN PASSED
SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE PASSED
SVGMarkerElement.SVG_MARKERUNITS_STROKEWIDTH PASSED
SVGMarkerElement.SVG_MARKER_ORIENT_UNKNOWN PASSED
SVGMarkerElement.SVG_MARKER_ORIENT_AUTO PASSED
SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE PASSED
SVGPathSeg.PATHSEG_UNKNOWN PASSED
SVGPathSeg.PATHSEG_CLOSEPATH PASSED
SVGPathSeg.PATHSEG_MOVETO_ABS PASSED
SVGPathSeg.PATHSEG_MOVETO_REL PASSED
SVGPathSeg.PATHSEG_LINETO_ABS PASSED
SVGPathSeg.PATHSEG_LINETO_REL PASSED
SVGPathSeg.PATHSEG_CURVETO_CUBIC_ABS PASSED
SVGPathSeg.PATHSEG_CURVETO_CUBIC_REL PASSED
SVGPathSeg.PATHSEG_CURVETO_QUADRATIC_ABS PASSED
SVGPathSeg.PATHSEG_CURVETO_QUADRATIC_REL PASSED
SVGPathSeg.PATHSEG_ARC_ABS PASSED
SVGPathSeg.PATHSEG_ARC_REL PASSED
SVGPathSeg.PATHSEG_LINETO_HORIZONTAL_ABS PASSED
SVGPathSeg.PATHSEG_LINETO_HORIZONTAL_REL PASSED
SVGPathSeg.PATHSEG_LINETO_VERTICAL_ABS PASSED
SVGPathSeg.PATHSEG_LINETO_VERTICAL_REL PASSED
SVGPathSeg.PATHSEG_CURVETO_CUBIC_SMOOTH_ABS PASSED
SVGPathSeg.PATHSEG_CURVETO_CUBIC_SMOOTH_REL PASSED
SVGPathSeg.PATHSEG_CURVETO_QUADRATIC_SMOOTH_ABS PASSED
SVGPathSeg.PATHSEG_CURVETO_QUADRATIC_SMOOTH_REL PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_UNKNOWN PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_NONE PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMINYMIN PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMIDYMIN PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMAXYMIN PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMINYMID PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMIDYMID PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMAXYMID PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMINYMAX PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMIDYMAX PASSED
SVGPreserveAspectRatio.SVG_PRESERVEASPECTRATIO_XMAXYMAX PASSED
SVGPreserveAspectRatio.SVG_MEETORSLICE_UNKNOWN PASSED
SVGPreserveAspectRatio.SVG_MEETORSLICE_MEET PASSED
SVGPreserveAspectRatio.SVG_MEETORSLICE_SLICE PASSED
SVGRenderingIntent.RENDERING_INTENT_UNKNOWN PASSED
SVGRenderingIntent.RENDERING_INTENT_AUTO PASSED
SVGRenderingIntent.RENDERING_INTENT_PERCEPTUAL PASSED
SVGRenderingIntent.RENDERING_INTENT_RELATIVE_COLORIMETRIC PASSED
SVGRenderingIntent.RENDERING_INTENT_SATURATION PASSED
SVGRenderingIntent.RENDERING_INTENT_ABSOLUTE_COLORIMETRIC PASSED
SVGTextContentElement.LENGTHADJUST_UNKNOWN PASSED
SVGTextContentElement.LENGTHADJUST_SPACING PASSED
SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS PASSED
SVGTextPathElement.TEXTPATH_METHODTYPE_UNKNOWN PASSED
SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN PASSED
SVGTextPathElement.TEXTPATH_METHODTYPE_STRETCH PASSED
SVGTextPathElement.TEXTPATH_SPACINGTYPE_UNKNOWN PASSED
SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO PASSED
SVGTextPathElement.TEXTPATH_SPACINGTYPE_EXACT PASSED
SVGTransform.SVG_TRANSFORM_UNKNOWN PASSED
SVGTransform.SVG_TRANSFORM_MATRIX PASSED
SVGTransform.SVG_TRANSFORM_TRANSLATE PASSED
SVGTransform.SVG_TRANSFORM_SCALE PASSED
SVGTransform.SVG_TRANSFORM_ROTATE PASSED
SVGTransform.SVG_TRANSFORM_SKEWX PASSED
SVGTransform.SVG_TRANSFORM_SKEWY PASSED
SVGUnitTypes.SVG_UNIT_TYPE_UNKNOWN PASSED
SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE PASSED
SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX PASSED
SVGViewElement.SVG_ZOOMANDPAN_UNKNOWN PASSED
SVGViewElement.SVG_ZOOMANDPAN_DISABLE PASSED
SVGViewElement.SVG_ZOOMANDPAN_MAGNIFY PASSED
